Support Fields API in conditional ingest processors
Fixing whitelist

Fix access to dotted field from source in yaml test

Separate field access and field modification APIs

Extending yaml test to cover all field access APIs

Separate WriteField and SourceMapField tests

Add test verifying that iterator() API does not allow removal

Small test change

Caching script instance instead of script factory

Update docs/changelog/121914.yaml

Initial review changes

Removing params from execute

Remove todo

Removing Field's methods from whitelists

[CI] Auto commit changes from spotless

Not relying on specific error type in yaml test
