+++ modules/lang-painless/src/main/resources/org/elasticsearch/painless/org.elasticsearch.script.ingest.txt
@@ -46,15 +46,12 @@ class org.elasticsearch.script.IngestScript {
 }
 
 class org.elasticsearch.script.field.WriteField {
-    String getName()
     boolean exists()
     WriteField move(def)
     WriteField overwrite(def)
     void remove()
     WriteField set(def)
     WriteField append(def)
-    boolean isEmpty()
-    int size()
     Iterator iterator()
     def get(def)
     def get(int, def)
