+++ server/src/main/java/org/elasticsearch/ingest/ConditionalProcessor.java
@@ -11,6 +11,7 @@ package org.elasticsearch.ingest;
 
 import org.elasticsearch.common.logging.DeprecationCategory;
 import org.elasticsearch.common.logging.DeprecationLogger;
+import org.elasticsearch.script.CtxMapWrapper;
 import org.elasticsearch.script.DynamicMap;
 import org.elasticsearch.script.IngestConditionalScript;
 import org.elasticsearch.script.Script;
@@ -57,6 +58,7 @@ public class ConditionalProcessor extends AbstractProcessor implements WrappingP
     private final IngestMetric metric;
     private final LongSupplier relativeTimeProvider;
     private final IngestConditionalScript precompiledConditionScript;
+    private final CtxMapWrapper ctxMapWrapper;
 
     ConditionalProcessor(String tag, String description, Script script, ScriptService scriptService, Processor processor) {
         this(tag, description, script, scriptService, processor, System::nanoTime);
@@ -76,11 +78,12 @@ public class ConditionalProcessor extends AbstractProcessor implements WrappingP
         this.processor = processor;
         this.metric = new IngestMetric();
         this.relativeTimeProvider = relativeTimeProvider;
+        this.ctxMapWrapper = new CtxMapWrapper();
 
         try {
             final IngestConditionalScript.Factory factory = scriptService.compile(script, IngestConditionalScript.CONTEXT);
             if (ScriptType.INLINE.equals(script.getType())) {
-                precompiledConditionScript = factory.newInstance(script.getParams());
+                precompiledConditionScript = factory.newInstance(script.getParams(), ctxMapWrapper);
             } else {
                 // stored script, so will have to compile at runtime
                 precompiledConditionScript = null;
@@ -144,9 +147,14 @@ public class ConditionalProcessor extends AbstractProcessor implements WrappingP
         IngestConditionalScript script = precompiledConditionScript;
         if (script == null) {
             IngestConditionalScript.Factory factory = scriptService.compile(condition, IngestConditionalScript.CONTEXT);
-            script = factory.newInstance(condition.getParams());
+            script = factory.newInstance(condition.getParams(), ctxMapWrapper);
         }
-        return script.execute(new UnmodifiableIngestData(new DynamicMap(ingestDocument.getSourceAndMetadata(), FUNCTIONS)));
+        ctxMapWrapper.setCtxMap(new UnmodifiableIngestData(new DynamicMap(ingestDocument.getSourceAndMetadata(), FUNCTIONS)));
+        try {
+            return script.execute();
+        } finally {
+            ctxMapWrapper.clearCtxMap();
+        }
     }
 
     public Processor getInnerProcessor() {
