+++ test/framework/src/main/java/org/elasticsearch/script/MockScriptEngine.java
@@ -169,10 +169,13 @@ public class MockScriptEngine implements ScriptEngine {
         } else if (context.instanceClazz.equals(AggregationScript.class)) {
             return context.factoryClazz.cast(new MockAggregationScript(script));
         } else if (context.instanceClazz.equals(IngestConditionalScript.class)) {
-            IngestConditionalScript.Factory factory = parameters -> new IngestConditionalScript(parameters) {
+            IngestConditionalScript.Factory factory = (parameters, ctxMapWrapper) -> new IngestConditionalScript(
+                parameters,
+                ctxMapWrapper
+            ) {
                 @Override
-                public boolean execute(Map<String, Object> ctx) {
-                    return (boolean) script.apply(ctx);
+                public boolean execute() {
+                    return (boolean) script.apply(ctxMapWrapper.getCtxMap());
                 }
             };
             return context.factoryClazz.cast(factory);
